requirements.txt
# Core HTTP & Config
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0

# AI Providers
openai==1.30.1
anthropic==0.25.7

# Scheduling
APScheduler==3.10.4

# Structured Logging
structlog==24.1.0

# Git integration
GitPython==3.1.43

# Linting & Testing (runtime validation)
flake8==7.0.0
pytest==8.2.0
pytest-mock==3.14.0

# Utilities
tenacity==8.3.0
click==8.1.7
rich==13.7.1
